refactor: move context to own file
